src/util/Memory: Moved the util to its own folder, in preparation for multi file organization.
